code-bot/requirements.txt
discord.py >= 1
python-decouple >= 3.1
twitchio >= 1